@@ -3,32 +3,26 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.7.0",
-    "@emotion/styled": "^11.6.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-brands-svg-icons": "^5.15.4",
-    "@fortawesome/free-regular-svg-icons": "^5.15.4",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.16",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@uiball/loaders": "^1.2.3",
-    "axios": "^0.24.0",
-    "framer-motion": "^4.1.17",
-    "i18next": "^21.6.16",
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
+    "@fortawesome/react-fontawesome": "^0.1.18",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^14.1.1",
+    "@uiball/loaders": "^1.2.6",
+    "axios": "^0.27.2",
+    "i18next": "^21.7.1",
     "i18next-browser-languagedetector": "^6.1.4",
     "i18next-http-backend": "^1.4.0",
-    "react": "^17.0.2",
+    "react": "^18.1.0",
     "react-country-flag": "^3.0.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.1.0",
     "react-ga": "^3.3.0",
-    "react-i18next": "^11.16.6",
-    "react-loader-spinner": "^4.0.0",
+    "react-i18next": "^11.16.9",
     "react-router": "^6.3.0",
     "react-router-dom": "^6.3.0",
     "react-router-sitemap": "^1.2.0",
-    "react-scripts": "^5.0.0"
+    "react-scripts": "^5.0.1"
   },
   "scripts": {
     "serve": "serve -s build",
@@ -58,12 +52,13 @@
     ]
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.2",
+    "autoprefixer": "^10.4.7",
     "babel-cli": "^6.26.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "babel-register": "^6.26.0",
-    "postcss": "^8.4.6",
-    "tailwindcss": "^3.0.18"
+    "postcss": "^8.4.13",
+    "tailwindcss": "^3.0.24",
+    "yarn-upgrade-all": "^0.7.1"
   }
 }
