@@ -1,22 +1,18 @@
-import React, { Suspense, lazy } from 'react';
-import ReactDOM from 'react-dom';
+import React from 'react';
+import { createRoot } from 'react-dom/client';
 import ReactGA from 'react-ga';
 import './index.css';
 import './styles/custom.css';
 import { I18nextProvider } from 'react-i18next';
 import i18n from './utils/i18n';
 import * as gtag from './utils/gtag';
+import App from './App';
 
 // Google Analytics
 ReactGA.initialize(gtag.GA_TRACKING_ID);
 
-const AppRoot = lazy(() => import('./App'));
-
-ReactDOM.render(
+createRoot(document.getElementById('root')).render(
     <I18nextProvider i18n={i18n}>
-        <Suspense fallback={<div className='text-2xl font-bold text-blurple text-center translate-y-1/2'>Loading...</div>}>
-            <AppRoot />
-        </Suspense>
-    </I18nextProvider>,
-    document.getElementById('root')
+        <App />
+    </I18nextProvider>
 );
